Move nationality autocomplete JS to its own file
